+++ docs/source/nlp/nlp_webserver_demo_processors.py
@@ -12,7 +12,7 @@ NLPRP web server. The PROCESSORS variable is the one of interest.
 from crate_anon.nlp_manager.all_processors import (
     all_crate_python_processors_nlprp_processor_info
 )
-from crate_anon.nlprp.constants import NlprpKeys as NKeys
+from crate_anon.nlprp.constants import NlprpValues, NlprpKeys as NKeys
 from crate_anon.nlp_webserver.constants import (
     KEY_PROCTYPE,
     PROCTYPE_GATE,
@@ -37,7 +37,8 @@ PROCESSORS = [
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "Finds mentions of drug prescriptions, "
                            "including the dose, route and frequency.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "diagnosis",
@@ -46,7 +47,8 @@ PROCESSORS = [
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "Finds mentions of diagnoses, in words or "
                            "in coded form.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "blood-pressure",
@@ -54,7 +56,8 @@ PROCESSORS = [
         NKeys.VERSION: "0.1",
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "Finds mentions of blood pressure measurements.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "cbt",
@@ -63,7 +66,8 @@ PROCESSORS = [
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "Identifies mentions of cases where the patient "
                            "has attended CBT sessions.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "lives-alone",
@@ -71,7 +75,8 @@ PROCESSORS = [
         NKeys.VERSION: "0.1",
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "Identifies if the patient lives alone.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "mmse",
@@ -82,7 +87,8 @@ PROCESSORS = [
                            "Extractor finds the results of this common "
                            "dementia screening test within documents along "
                            "with the date on which the test was administered.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "bmi",
@@ -90,7 +96,8 @@ PROCESSORS = [
         NKeys.VERSION: "0.1",
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "Finds mentions of BMI scores.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "smoking",
@@ -100,7 +107,8 @@ PROCESSORS = [
         NKeys.DESCRIPTION: "Identifies instances of smoking being discussed "
                            "and determines the status and subject (patient or "
                            "someone else).",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "ADR",
@@ -108,7 +116,8 @@ PROCESSORS = [
         NKeys.VERSION: "0.1",
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "Adverse drug event mentions in clinical notes.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "suicide",
@@ -116,7 +125,8 @@ PROCESSORS = [
         NKeys.VERSION: "0.1",
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "App derived from TextHunter project suicide.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "appetite",
@@ -124,7 +134,8 @@ PROCESSORS = [
         NKeys.VERSION: "0.1",
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "Finds markers of good or poor appetite.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
     {
         NKeys.NAME: "low_mood",
@@ -132,7 +143,8 @@ PROCESSORS = [
         NKeys.VERSION: "0.1",
         NKeys.IS_DEFAULT_VERSION: True,
         NKeys.DESCRIPTION: "App derived from TextHunter project low_mood.",
-        KEY_PROCTYPE: PROCTYPE_GATE
+        KEY_PROCTYPE: PROCTYPE_GATE,
+        NKeys.SCHEMA_TYPE: NlprpValues.UNKNOWN
     },
 ] + all_crate_python_processors_nlprp_processor_info()
 
@@ -147,4 +159,4 @@ if __name__ == "__main__":
 
 
 
-# Generated at 2019-08-22 16:29:34
+# Generated at 2019-09-01 22:19:26